Read session tokens from the storage
