+++ web_patient/src/components/Chrome/AppHost.tsx
@@ -9,6 +9,7 @@ import { useServices } from 'src/services/services';
 import { IRootStore, RootStore } from 'src/stores/RootStore';
 import { StoreProvider } from 'src/stores/stores';
 import styled, { withTheme } from 'styled-components';
+import AppLoader from 'src/components/Chrome/AppLoader';
 
 const LoginContainer = withTheme(
     styled.div((props) => ({
@@ -84,6 +85,8 @@ export const AppHost: FunctionComponent<IAppHost> = observer((props) => {
                 runInAction(() => {
                     state.store = newStore;
                 });
+
+                newStore.authStore.initialize();
             })
             .catch((error) => {
                 console.error('Failed to retrieve server configuration', error);
@@ -117,6 +120,8 @@ export const AppHost: FunctionComponent<IAppHost> = observer((props) => {
                     <StoreProvider store={state.store}>{children}</StoreProvider>
                 </Fragment>
             )}
+
+            <AppLoader isLoading={!!state.store?.authStore.isAuthenticating} text="Logging in..." />
             <Fade in={!state.store?.authStore.isAuthenticated} timeout={500}>
                 <LoginContainer>
                     {!!state.store ? (
