+++ web_shared/package.json
@@ -2,6 +2,7 @@
     "name": "scope-web-shared",
     "@comment dependencies": [],
     "dependencies": {
+        "amazon-cognito-identity-js": "^5.2.7",
         "axios": "^0.21.1",
         "date-fns": "^2.17.0",
         "date-fns-tz": "^1.2.2",
